Fix: bug when running the main scene
The tendrilis didn't use the show_factor variable to display the text.
The problem was using the export variable with the setget functions with
"if not is_node_ready(): await ready".
The correct way is to init everything in the _ready function and in the
setget functions use "if is_node_ready():" do what should be done.